+++ packages/shared/src/server/services/PromptService/utils.ts
@@ -1,0 +1,3 @@
+export function escapeRegex(str: string | number) {
+  return String(str).replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
+}
